Don't delete proctor keys for tests with unlimited number of attempts
Currently if a proctored test is set to have an unlimited number of
attempts per version, the user clicks to grade the test, and the proctor
credentials are validly entered, then if the user tries to move to
another page of the test the user is taken to the proctor login page
again.  This is because of an error in the logic that causes the proctor
keys to be deleted instead of the proctor grading key.