Switch to using a Mojolicious session for cookies.
The primary advantage of this switch is that now cookies are signed.
That means that unlike webwork2's current cookies, the data stored in
the session cookies can be trusted to not have been tampered with. Note
that setting the secrets in the `conf/webwork2.mojolicious.yml` file is
now important (and not just something that we tell system administrators
to do).  That secret is the private key that is used to sign cookies.

The documentation on session management in defaults.config and
localOverrides.conf.dist has been updated.  Most of that was obsolete
nonsense.

Note that the $debugPracticeUser has been removed.  As stated in the
comments about it in defaults.config, "Why de we even have this?"