Tweaks for the PG problem editor for course info files.
This wraps the injected course info html in a `div` that has overflow
auto set.  If the course info file has long contents, then those
contents currently extend past the containing div, and this fixes that.

In addition padding and a background color are added to improve the
appearance.  These are the same styles that are used when this is
displayed in the actual info box.