Fix how invalid problem ids from set definition files are handled.
This identifies the maximum valid problem id, and then invalid problem
ids that are found use one more than that.  Of course the max is
incremented for later invalid problem ids.  This is the simplest
approach.

So this gives the following numberings:
set def  ->  result
`0,1,2,3` -> `1,2,3,4`
`0,2,3,4` -> `2,3,4,5`
`0,1,2,5` -> `1,2,5,6`

Another option that could be implemented is to fill when gaps in the
valid numbers are found.  That takes a bit more work, but could be done.