Implement suggestions made in my review of openwebwork/webwork2#2266
This uses `flex` rules to make the category input groups behave better.
The categories are wrapped in a div with fixed width.  The width is
chosen to be as small as possible such that the input groups do not
wrap.  Thus this dispenses with trying to make the category widths match
the scrolling record list which really wasn't working.  `flex-wrap` is
used with "around" justification.  So at extra large window widths all
three categories will be in the same row.  However, when the categories
don't fit, they wrap.  The justification gives nicer positioning in any
case.

I merged the input texts, removed capitalization, and changed some of
the wording in a few cases.  For example, I don't think that the
"Library Browser Exercises" is good.  I think the reference to the
library browser is unnecessary, and using "exercises" instead of
"problems" is inconsistent the wording used everywhere else in webwork2.

The other changes made are more explicit in the comments made in my
review.