Add problem seed to past answers.
Add a `problem_seed` column to the `past_answer` database table.  This
is displayed in the first column of the table shown on the "Answer Log"
page.  Since that table is not a fully qualified data table and is
missing headers, it is displayed in the table as "Seed: 123".  If a
problem does not have a seed (because it is from a past answer before
this pull request like an archived course), then "Seed: unknown" is
shown.