Fix filtering the user list by permission level or status for other languages.
Currently when filtering by permission level or status the text entered
in the "Filter by what text?" field is matched against the untranslated
permission level name or status name.  As a result other languages must
enter the English names for these to match what is displayed translated
which is probably very confusing.  This now matches against the
translated names as well as the untranslated names.

This fixes issue #1093.