Style tweaks for the new tabbed layout of the instructor tools input groups.
At different window widths there is now a lot of wrapping of the input
groups since we have switched back to using the bootstrap grid layout.
Particularly with the latest wording changes that make some of the input
group texts even longer.  So this adjusts the `col` sizes at narrower
window widths to prevent that, and allow ample room for translations
that might be longer.

Make the pill buttons match the width of the scrolling record list above
at all layouts, and add a small margin to give them some separation.

Add a missing `row` div to contain the `instructor-tools-nav`.  In
bootstrap's grid layout a `col` must be a direct descendent of a `row`
div.

Remove the `col-2` on the submit buttons.  Those are not direct
descendents of a `row` div, and so should not have that class.
Furthermore, the `min-width` set in math4.scss overrides that anyway.

Make the margin below the scrolling record list match the margin after
the pill buttons for even spacing.

Fix some overly long lines.