Compile the achievement notification email templates in a safe compartment.
To do so the WeBWorK::SafeTemplate module must be used.  This module
derives from the Mojo::Template module and overrides the _trap method.
The Mojo::Template _wrap method adds "use Mojo::Base -strict" and "no
warnings 'ambiguous'" to the code generated in this method.  When that
is called later it causes an error in the safe compartment because
"require" is trapped. So instead the WeBWorK::SafeTemplate override
method imports strict, warnings, and utf8 which is equivalent to calling
"use Mojo::Base -strict". Calling "no warnings 'ambiguous'" prevents
warnings from a lack of parentheses on a scalar call in the generated
code.  So if this package is used, those warnings need to be prevented
in another way.  That is done when the module is used in
Mojolicious::WeBWorK::Tasks::AchievementNotification using a
$SIG{__WARN__} handler that doesn't do anything if the ambiguous warning
is encountered.