Improvements to the essay "needs_grading" flag.
First, the single problem grader now removes this flag when an essay
answer is graded.  A new `mark_graded` parameter for the
WebworkWebservice::ProblemActions `putUserProblem` and
`putProblemVersion` methods is added to accomodate this.

When a problem is rendered this flag now needs to be passed to PG to
indicated whether a problem still needs to be graded so that essay
answers can not always show "Ungraded".

The `create_ans_str_from_responses` method now handles adding this flag
instead of that code being duplicated in ProblemProcessing.pm and
GatewayQuiz.pm.