Suggestions for improvement.
Inline javascript is bad form.  There is already a javascript file
dedicated to the ProblemSetDetail.pm page, so put the javascript there
instead.

The return value of the Mojolicious::Plugin::TagHelpers `tag` method is
already a Mojo::ByteStream object.  The issue that causes the display to
not work is that the contents of the `div` are not, since they are just
bare strings concatentated together, and so they are html escaped.  So
use a Mojo::Collection that is joined by the empty string.  That also
returns a Mojo::ByteStream object, and so it is not html escaped. This
is also consistent with how I have done this sort of thing elsewhere in
the code.

Also remove what appears to be an errant `warn` statement.