Switch to using Github links for bug reporting.
The site nav bug link now links to
https://github.com/openwebwork/webwork2/issues/new.  This link does not
add a title or body for the issue.

The PG problem editor bug link now links to
https://github.com/openwebwork/webwork-open-problem-library/issues/new.
This link sets the issue title to the problem file path (relative to the
templates directory of the course), and adds the current problem seed to
the issue body.