Fix problem ids displayed for achievement items.
Instead of just storing a max problem id for each achievement set
option, store the json encoded list of problem ids.  Initially only the
options for the first set are provided.  The javascript then rebuilds
the options list when a new set is selected that only includes the
problem ids for that set.

This also restructures the way that achievement item data is obtained
from the database to be a bit more efficient.  When this data is
obtained gateway tests are no longer filtered out.  The current code for
each achievement item that does not act on gateway tests is already set
up to filter these out.  The achievement items that do act on gateway
tests no longer need to load data from the database on their own.  They
can use the given list and filter out non-gateway sets.

There seems to have been a minor bug in the `ResurrectGW` achievement in
that it showed all gateway sets to be possibly "resurrected" when it
should only show gateway sets that are past due.  That has also been
fixed, so now it only shows past due gateway sets for "resurrection".

This fixes issue #2281.