Make achievements work on the unreduced score.
Achievements now compute and use the unreduced score for problems for
determining if a problem is complete/correct.  That is set as the
`status` that is passed into the achievement evaluators.  The reduced
score is cached in a hash key of the `$problem` object named
reduced_score in case an achievement evaluator wants that information
also (although none of them use this at this point).

This fixes issue #2285.