Make the set progress bar work with unreduced scores for determining correctness/completion of a problem.
The siblings problem list shows a checkmark by problems for which the
maximum possible score has been obtained, even if that is not 100% for
the problem due to being in the reduced scoring period.  Also when all
problems are complete in the set and there are problems with reduced
scores, then the progress bar shows "Complete" without stars instead of
the usual "Perfect" with stars.

Also mark "Perfect" for translation.  For some reason it was not before.