Clean up of the htdocs/themes directories.
Move gateway.scss and achievements.scss into htdocs/js subfolders.
Rename the math4.js file to htdocs/js/System/system.js.  Rename the
math4.scss file to htdocs/js/System/system.scss.  These files do not
belong in the theme directories and should not be considered theme-able.

Where possible, the page specific code in the math4.js (now system.js)
file has been moved into the page specific JavaScript file for that
page.

The aria label hack that was in math4.js has been removed.  It is clear
that is not doing the correct thing, and it is probably less of a
problem to have an aria label that reads "answer 1" even when there are
no other answers than the numerous cases that it is changing aria labels
to just "answer" when it shouldn't.  One case is if there is only one
text answer and there are other answers (radio buttons, checkboxes, etc)
on the page.  In this case the "answer 1" is changed to "answer" even
when there is more than one answer on the page.  Also in gateway quizzes
the "problem n" prefix is removed if there is only one problem on the
page, and it shouldn't be.  Aria labels in general need to be rethought,
but this JavaScript hack is just incorrect.

Update the README files in the themes directories a bit.  More could be
done here to give better directions on how to customize themes.

The name of the htdocs/js directory really needs to be changed.  It is
really the public "assets" location, and not just JavaScript.  The
htdocs/css directory should be removed as well.

Also change "npm install" to "npm ci" in the warnings that are issued if
the `static-assets.json` files are not found.