Move the system.html.ep template out of the themes.
That file is now the system layout instead.  This file should really not
be themable.  If someone wants to mess with that file, they can deal
with having locally modified files that show up as changed by git.  If
someone does not know what they are doing, they can really mess up
things.

If there is some other thing that is commonly wanted in the default
system layout it can be added as an option.

So now the themes directories only have what is needed for theming, and
theming for the most part is only about changing the color scheme.  This
is how it should be.  Theming should not mean modify the entire system
to make it be what you want.

There are several advantages with this pull request.  It removes the
need for so many symlinks from the math4 theme variants back to files in
the math4 theme.  It removes the need for the hackish
`htdocs/themes/layouts` symlink that turns the themes directory into a
layout directory.  It removes the need for the themes directory to be a
template path to begin with.