Add the AuthorComment css back once again.
I keep forgetting why this is needed both here and in PG (actually it
technically is not needed for PG), and so delete it here.  The issue is
that author comments are rendered on pages (the PG problem editor and
library browser) where the problems are rendered in iframes, and the
comment is outside of the iframe.  So the problem.scss styles of PG only
apply to what is inside the iframe, and this style is needed here to
apply to the author comments not in the iframe.

To prevent me from deleting this again (or anyone else), I added a
comment this time to remind me.