Fix instructor feedback not being given when showPartialCorrectAnswers is false.
This applies to problems rendered in the library browser or PG problem
editor.  In that case if $showPartialCorrectAnswers is false, then the
instructor is not given the problem status (correct/incorrect) in the
feedback.  This is because the renderProblem method does not set the
forceShowAttemptResults (which overrides the $showPartialCorrectAnswers
variable in PG) option.  So that is fixed by setting that option from
the submit buttons for instructors.