Add content item selection for LTI.
See https://www.imsglobal.org/specs/lticiv1p0/specification
and https://www.imsglobal.org/spec/lti-dl/v2p0.

This will be needed for Moodle 4.03 (and beyond?) but might be nice for
other LMS's as well.  If the new content selection endpoints
(https://your.webwork2.server.domain/ltiadvanced/content_selection for
LTI 1.1 and
https://your.webwork2.server.domain/ltiadvantage/content_selection for
LTI 1.3) are added to the LMS external tool configuration, then
instructors can select from a list of homework assignments in the
webwork2 course, and it will add the correct link(s) to the LMS.  For
Moodle (at least for 4.1 and up) you can add links to all assignments
with a few clicks of the mouse.  Canvas is not quite as nice as it only
allows selection one assignment at a time (at least as far as I have
been able to figure out at this point).  As usual, I am not sure what
D2L can do.

In order for this to work each webwork2 course must set the new LTI
configuration variable `$LMSCourseName` in its course.conf file.  This
is the name of the LMS course that is associated with the webwork2
course.  This must be done because these content item requests do not
have the webwork2 course name in the link like the usual webwork2
course or assignment links.

In order to support multiple LMS's with a single webwork2 server there
is a little more that must be done for LTI 1.1, particularly if there is
the possibility that courses from different LMS's might have the same
LMS course name.  In this case each LMS must use a different consumer
key for its external tool for the webwork2 server, and the
`$LTI{v1p1}{ConsumerKey}` must be set in each courses course.conf file.
For LTI 1.3 no additional configuration is needed because the
authentication parameters already uniquely identify an LMS external
tool.

Note that none of this new configuration is needed in the case that the
webwork2 server administrator does not want instructors to be able to
utilize the new content item selection routes.  The `$LTICourseName`
does not need to be set in the `course.conf` files, the
`$LTI{v1p1}{ConsumerKey}` does not need to be set, and both versions of
LTI will continue to work as they did before.

Now for implementation details.

For LTI 1.1 nothing is changed in the way that authentication works
unless the LMS sends the user to the new `ltiadvanced_content_selection`
route.  In that case the correct course is identified by iterating
through the webwork2 courses on the server, loading the course
environment for each one, and finding the one with the correct LMS
course name (and ConsumerKey if there is more than one with the same LMS
course name).  This is probably rather inefficient, but fortunately this
is only done for content item selection requests.

For LTI 1.3 things needed to be reworked quite a bit more.  All of the
hackery that was previously implemented to pass the course id, state,
and nonce from the login request to the launch request using the key
table had to go.  Instead there is now a non-native table (the
lti_launch_data table) that this information is stored in, identified by
the state generated from the parameters provided by the LMS in the login
request which are unique to the LMS user and request.  This table can be
accessed without a webwork2 course id since it is non-native.    This
allows for the state to be retrieved from the database in the launch
request, and the JWT to be decoded earlier in that request (previously
this was done in the authentication step, and now it is done before that
in the webwork2 dispatch step).  It needs to be done at dispatch time
because the JWT contains the information that determines if it is a deep
linking request or not.  For both the login and launch requests a
webwork2 course is identified by iterating through courses, loading
course environments, and finding an LTI 1.3 tool configuration that
matches that in the request.  Again this is probably rather inefficient,
but again this is only done for content item selection requests.

There is a new hook that can be utilized by content generator modules to
modify parameters, route captures, and do other things early in the
dispatch phase before a course id has been determined and a course
environment and database instance is initialized. The hook can be
utilized by the content generator module providing an `initializeRoute`
method.  This is just clean up really.  A lot of route specific special
cases were starting to build up in the `dispatch` method in
`lib/WeBWork.pm`, and this makes it possible for that code to be moved
out of there.