Update the `bin/addcourse` script for the new `addCourse` options.
This just uses the simplest possible approach and switches the script
from using the now removed `templatesFrom` option to using the new
`copyFrom` option.  If that option is set, then the new
`copyTemplatesHtml` option is also set to be true.  Thus the existing
behavior of the script is preserved.

The script could be updated to actually take advantage of the new
options instead if one were ambitious, but this is good enough for now.

This also fixes all of the long lines in
`lib/WeBWorK/Utils/CourseManagement.pm`, and an issue with comments in
the `initNonNativeTables` method.