Add two factor authentication.
This uses the TOTP (time-based one-time password) protocol. Any
authenticator app on a mobile device that supports this protocol can be
used (for example, Google Authenticator, Microsoft Authenticator, Twilio
Authy, etc.).

Whether this is enabled or not is controlled by new course environment
variable `$twoFA{enabled}`.  If that is set to 0, then two factor
authentication is disabled for all courses. If that is 1 (the default),
then two factor authentication is enabled for all courses that use
password authentication (of course two factor authentication does not
apply to courses that use external authentication methods like LTI, CAS,
Shibboleth, etc.). If that is a string course name like 'admin', then
two factor authentication is enabled only for that course. If that is an
array of string course names, then two factor authentication is enabled
only for those courses listed. This can also be set in a course's
course.conf file. Note that only the values of 0 and 1 make sense there.

There are two methods that can be used to setup two factor
authentication when a user signs in for the first time. The setup
information can be emailed to the user, or can be directly displayed in
the browser on the next page that is shown after password verification
succeeds.

This is controlled by the new course environment variable
`$twoFA{email_sender}`.  If that is set, then the email approach will be
used.  In this case, after a user signs in and the password is verified,
the user will be sent an email containing a QR code and instructions on
how to set up a OTP generator app. This is probably a more secure way to
set up two factor authentication, as it ensures the user setting it up
is the correct user. Note that if a user does not have an email address,
then the browser method below will be used as a fallback.

If `$twoFA{email_sender}` is not set, then after a user signs in and the
password is verified, the QR code, OTP link, and instructions will be
displayed directly on the page in the browser. This is potentially less secure
because a hacker could guess a username and password before a user has setup
two factor authentication (particularly if the username and password are
initially the same), and then the hacker would gain access to that user's
account, and the actual user would be locked out. Note that you will need to
use this option if your server can not send emails. Also note that no-reply
addresses may be blocked by the email server or marked as spam. So it may be
better to find a valid email address to use for this.

This requires a new database column `otp_secret` that was added to the
password table (for lack of a better place to put it).

There is a new wwsh script (`bin/reset2fa`) that can be used to reset
two factor authentication for a user if a user somehow loses their setup
in an authentactor app on their mobile device. That just removes the OTP
secret from the database.  This means that the user will need to go
through the two factor authentication setup process again.  To use it
execute: `wwsh courseID /opt/webwork/webwork2/bin/reset2fa userID`.
Multiple user ids can be listed if you want to reset more than one user
at a time. This is the only way that an admin user can reset their own
two factor authentication (and there should never be another way for
admin users to do this for themselves).  Perhaps a page in the admin
course could be added for resetting this for instructors. I plan to add
a way that instructors can reset two factor authentication for students.